libcamera: v4l2_device: Use `bool` for boolean controls
The `ControlInfo` generated by `V4L2Device::v4l2ControlInfo()` already
uses the `bool` type, and there is no reason to disallow setting proper
"bool" values.

So adjust `updateControls()` and `setControls()` accordingly, and also
do the necessary adjustments wrt. `V4L2_CID_{V,H}FLIP`.

+++ src/libcamera/sensor/camera_sensor_raw.cpp
@@ -822,10 +822,8 @@ int CameraSensorRaw::setFormat(V4L2SubdeviceFormat *format, Transform transform)
 	if (supportFlips_) {
 		ControlList flipCtrls(subdev_->controls());
 
-		flipCtrls.set(V4L2_CID_HFLIP,
-			      static_cast<int32_t>(!!(transform & Transform::HFlip)));
-		flipCtrls.set(V4L2_CID_VFLIP,
-			      static_cast<int32_t>(!!(transform & Transform::VFlip)));
+		flipCtrls.set(V4L2_CID_HFLIP, !!(transform & Transform::HFlip));
+		flipCtrls.set(V4L2_CID_VFLIP, !!(transform & Transform::VFlip));
 
 		int ret = subdev_->setControls(&flipCtrls);
 		if (ret)
